fixing sevrices

+++ src/controllers/projectController.js
@@ -1,12 +1,13 @@
 const { v4: uuidv4, validate: uuidValidate } = require('uuid')
 const projectService = require('../services/projectService')
+const userService = require('../services/userService')
 const { response } = require('../utils/middleware')
 
 const getProjects = async (req, res, next) => {
   try {
     const projects = await projectService.findAllProjects()
     if (projects.length === 0) {
-      return response(res, 404, 'No Projects found')
+      return response(res, 404, 'No Project found')
     }
     response(res, 200, 'All projects retrieved successfully', { projects })
   } catch (error) {
@@ -83,22 +84,48 @@ const deleteProject = async (req, res, next) => {
 
 const getProjectUsers = async (req, res, next) => {
   try {
-    const project = await projectService.findProjectUsers(req.params.id)
-    if (!project) {
-      return response(res, 404, `Project with ID: ${req.params.id} was not found`)
+    if (!uuidValidate(req.params.id)) {
+      return response(res, 404, `Project with ID: ${req.params.id} not found`);
     }
-    response(res, 200, 'Projects retrieved successfully', { project })
+    const project = await projectService.findProjectUsers(req.params.id);
+    if (!project) {
+      return response(res, 404, `Project with ID: ${req.params.id} was not found`);
+    }
+    if (!project.Users || project.Users.length === 0) {
+      return response(res, 404, `No users found for ${project.name}`);
+    }
+    response(res, 200, 'Project retrieved successfully', { project });
   } catch (error) {
-    response(res, 500, 'Internal Server Error', { error: error.message })
-    console.log(error)
-    next(error)
+    response(res, 500, 'Internal Server Error', { error: error.message });
+    console.log(error);
+    next(error);
   }
-}
+};
+
 const addUserToProject = async (req, res, next) => {
   try {
-    const { projectId, userId } = req.params
-    const project = await projectService.addUserToProject(projectId, userId)
-    response(res, 200, 'User added to project successfully', { project })
+    const { projectId, userId } = req.params;
+    if (!uuidValidate(projectId)) {
+      return response(res, 404, `Project with ID: ${projectId} not found`)
+    }
+    if (!uuidValidate(userId)) {
+      return response(res, 404, `User with ID: ${userId} not found`)
+    }
+    const project = await projectService.findProjectById(projectId)
+    const user = await userService.findUserById(userId)
+    if (!project) {
+      return response(res, 404, `Project with ID: ${projectId} not found or not created`)
+    }
+    if (!user) {
+      return response(res, 404, `User with ID: ${userId} not found or not registered`)
+    }
+    const isUserAdded = await projectService.isUserInProject(projectId, userId);
+    if (isUserAdded) {
+      return response(res, 400, `${user.fullName} is already added to the project`)
+    }
+    const ProjectUserData = { projectId, userId, id: uuidv4() }
+    const addProjectUser = await projectService.addUserToProject(ProjectUserData)
+    response(res, 200, 'User added to project successfully', { addProjectUser })
   } catch (error) {
     response(res, 500, 'Internal Server Error', { error: error.message })
     console.log(error)
@@ -109,6 +136,17 @@ const addUserToProject = async (req, res, next) => {
 const removeUserFromProject = async (req, res, next) => {
   try {
     const { projectId, userId } = req.params
+    if (!uuidValidate(projectId)) {
+      return response(res, 404, `Project with ID: ${projectId} not found`)
+    }
+    if (!uuidValidate(userId)) {
+      return response(res, 404, `User with ID: ${userId} not found`)
+    }
+    const user = await userService.findUserById(userId)
+    const isUserAdded = await projectService.isUserInProject(projectId, userId);
+    if (!isUserAdded) {
+      return response(res, 404, `${user.fullName} not in this project`)
+    }
     const project = await projectService.removeUserFromProject(projectId, userId)
     response(res, 200, 'User removed from project successfully', { project })
   } catch (error) {
